refactored packages updated demo
