+++ app/src/main/AndroidManifest.xml
@@ -9,13 +9,18 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".DemoActivity">
+        <activity
+            android:name=".DemoListActivity"
+            android:label="Demo List">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".DemoChatActivity"
+            android:label="Demo Chat" />
     </application>
 
 </manifest>
