+++ app/src/main/res/layout/activity_demo.xml
@@ -1,12 +1,20 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@android:color/white"
     android:orientation="vertical"
     android:paddingTop="10dp"
-    tools:context="in.habel.android_binding_adapters.DemoActivity">
+    tools:context="in.habel.android_binding_adapters.DemoChatActivity">
+
+    <android.support.v7.widget.SearchView
+        android:id="@+id/searchView"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        app:defaultQueryHint="Search here" />
+
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/recyclerView"
