+++ app/src/main/res/menu/home_menu.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <item
+        android:id="@+id/toggle"
+        android:icon="@mipmap/ic_launcher_round"
+        android:title="Switch"
+        android:visible="true"
+        app:showAsAction="always" />
+</menu>
