+++ chat-adapters/src/main/java/in/habel/animator/ViewHelper.java
@@ -1,4 +1,4 @@
-package in.habel.chat_adapters.animator;
+package in.habel.animator;
 
 import android.support.v4.view.ViewCompat;
 import android.view.View;
