+++ chat-adapters/src/main/java/in/habel/chat_adapters/RecyclerAdapter.java
@@ -1,4 +1,4 @@
-package in.habel.chat_adapters.baseadapter;
+package in.habel.chat_adapters;
 
 import android.databinding.DataBindingUtil;
 import android.databinding.ViewDataBinding;
@@ -14,90 +14,52 @@ import android.view.animation.OvershootInterpolator;
 
 import java.util.ArrayList;
 
-import in.habel.chat_adapters.animator.FadeInUpAnimator;
-import in.habel.chat_adapters.interfaces.RecyclerChatCallback;
-import in.habel.chat_adapters.interfaces.chatInterface;
+import in.habel.animator.FadeInUpAnimator;
+import in.habel.interfaces.RecyclerCallback;
 
-/**
- * @param <T>  Chat data model
- * @param <VM> DataBinding class of incoming messages
- * @param <VN> DataBinding class of outgoing messages
- */
-@SuppressWarnings("unused")
-public class RecyclerChatAdapter<T extends chatInterface, VM extends ViewDataBinding, VN extends ViewDataBinding> extends RecyclerView.Adapter<RecyclerChatAdapter.ViewHolder> {
-    private final RecyclerView recyclerView;
-    private LinearLayoutManager linearLayoutManager;
-    private ArrayList<T> items;
-    private int incomingLayoutId, outgoingLayoutId;
+public class RecyclerAdapter<T, VM extends ViewDataBinding> extends RecyclerView.Adapter<RecyclerAdapter.ViewHolder> {
+    private RecyclerView recyclerView;
+    private int layoutId;
     private int maxSeenPosition;
+    private LinearLayoutManager linearLayoutManager;
+    @Nullable
+    private RecyclerCallback<VM, T> bindingInterface;
+    private ArrayList<T> items;
     private boolean scrollToLast;
 
-    @Nullable
-    private RecyclerChatCallback<VM, VN, T> bindingInterface;
-
-    /**
-     * Creates a recycler chat adapter.
-     *
-     * @param recyclerView     RecyclerView to which this adapter binds
-     * @param items            dataset of <T>
-     * @param incomingLayoutId view resource id for incoming messages
-     * @param outgoingLayoutId view resource id for outgoing messages
-     * @param bindingInterface Callback for binding adapter
-     */
-    public RecyclerChatAdapter(RecyclerView recyclerView, ArrayList<T> items, int incomingLayoutId, int outgoingLayoutId, @Nullable RecyclerChatCallback<VM, VN, T> bindingInterface) throws Exception {
+    public RecyclerAdapter(RecyclerView view, ArrayList<T> items, int layoutId, @Nullable RecyclerCallback<VM, T> bindingInterface) {
         this.items = items;
-        if (recyclerView == null)
-            throw new Exception("Recycler view for RecyclerChatAdapter cannot be null");
-        this.recyclerView = recyclerView;
-        this.incomingLayoutId = incomingLayoutId;
-        this.outgoingLayoutId = outgoingLayoutId;
+        recyclerView = view;
+        this.layoutId = layoutId;
         this.bindingInterface = bindingInterface;
         linearLayoutManager = new LinearLayoutManager(recyclerView.getContext());
-        linearLayoutManager.setSmoothScrollbarEnabled(true);
         recyclerView.setLayoutManager(linearLayoutManager);
-        linearLayoutManager.setStackFromEnd(true);
         recyclerView.setAdapter(this);
         FadeInUpAnimator animator = new FadeInUpAnimator(new OvershootInterpolator(.2f));
         recyclerView.setItemAnimator(animator);
-        //  maxSeenPosition = items.size();
     }
 
     @Override
     public ViewHolder onCreateViewHolder(ViewGroup parent,
                                          int viewType) {
-        boolean out = items.get(viewType).isOutgoing();
         View v = LayoutInflater.from(parent.getContext())
-                .inflate(out ? outgoingLayoutId : incomingLayoutId, parent, false);
-        return new ViewHolder(v, !out);
+                .inflate(layoutId, parent, false);
+        return new ViewHolder(v);
     }
 
 
     @SuppressWarnings("unchecked")
     @Override
-    public void onBindViewHolder(RecyclerChatAdapter.ViewHolder holder, int position) {
+    public void onBindViewHolder(RecyclerAdapter.ViewHolder holder, int position) {
         T item = items.get(position);
         holder.bindData(item);
         maxSeenPosition = Math.max(maxSeenPosition, linearLayoutManager.findLastVisibleItemPosition());
     }
 
-
-    /**
-     * @return size of dataset
-     */
     @Override
     public int getItemCount() {
         return items == null ? 0 : items.size();
     }
-
-    public RecyclerView getRecyclerView() {
-        return recyclerView;
-    }
-
-    @Override
-    public int getItemViewType(int position) {
-        return position;
-    }
-
     // Insert a new item to the RecyclerView on a predefined position
     public void insert(T data) {
         int position = items.size();
@@ -123,16 +85,13 @@ public class RecyclerChatAdapter<T extends chatInterface, VM extends ViewDataBin
         try {
             maxSeenPosition = Math.max(maxSeenPosition, linearLayoutManager.findLastVisibleItemPosition());
             if (bindingInterface != null) {
-                //   bindingInterface.onUnreadMessageFound(getItemCount(), items.size() - 2 - maxSeenPosition);
+                bindingInterface.onUnreadMessageFound(getItemCount(), items.size() - 2 - maxSeenPosition);
             }
         } catch (Exception e) {
             e.printStackTrace();
         }
     }
 
-    public void setScrollToBottom(boolean scrollToBottom) {
-        scrollToLast = scrollToBottom;
-    }
     private void scrollIfLast() {
         if (!scrollToLast) return;
         int lastVisiblePosition = linearLayoutManager.findLastVisibleItemPosition();
@@ -147,6 +106,9 @@ public class RecyclerChatAdapter<T extends chatInterface, VM extends ViewDataBin
         }
     }
 
+    public void setScrollToBottom(boolean scrollToBottom) {
+        scrollToLast = scrollToBottom;
+    }
 
     // Remove a RecyclerView item containing a specified Data object
     public void remove(T data) {
@@ -160,13 +122,11 @@ public class RecyclerChatAdapter<T extends chatInterface, VM extends ViewDataBin
         scrollIfLast();
     }
 
-
     @Deprecated
     public void refresh() {
         notifyDataSetChanged();
         scrollIfLast();
     }
-
 
     @SuppressWarnings("unchecked")
     public synchronized void refresh(ArrayList<T> newData) {
@@ -193,24 +153,19 @@ public class RecyclerChatAdapter<T extends chatInterface, VM extends ViewDataBin
         for (int i = newSize, itemSize = items.size(); i < itemSize; i++) remove(newSize);
     }
 
+
     public class ViewHolder extends RecyclerView.ViewHolder {
 
-        @Nullable
-        VM incomingBinding;
-        @Nullable
-        VN outgoingBinding;
+        VM binding;
 
-        public ViewHolder(View view, boolean incoming) {
+        public ViewHolder(View view) {
             super(view);
-            if (incoming)
-                incomingBinding = DataBindingUtil.bind(view);
-            else
-                outgoingBinding = DataBindingUtil.bind(view);
+            binding = DataBindingUtil.bind(view);
         }
 
         void bindData(T model) {
             if (bindingInterface != null)
-                bindingInterface.bindData(incomingBinding, outgoingBinding, model);
+                bindingInterface.bindData(binding, model);
         }
 
     }
