+++ chat-adapters/src/main/java/in/habel/interfaces/chatInterface.java
@@ -1,4 +1,4 @@
-package in.habel.chat_adapters.interfaces;
+package in.habel.interfaces;
 
 /**
  * Created by habel on 23/4/17.
